TermsTrie and SuffixTrie to separate modules
